% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_alt.R
\name{create_map_alt}
\alias{create_map_alt}
\title{Create Summary Map of Sampling Locations}
\usage{
create_map_alt(
  .data,
  type = "circle",
  parameter = NA,
  parameterColumn = "CharacteristicName",
  digits = 3
)
}
\arguments{
\item{.data}{full WQP data frame with site location data}

\item{type}{type of map to generate (i.e. simple map or circle marker map)}

\item{parameter}{optional input if user wants to map mean values for a parameter (by default plots all available parameters)}

\item{parameterColumn}{name of the column with parameter names (not likely to require adjustment)}

\item{digits}{digits used for displaying rounded data}
}
\value{
an interactive map of sample locations displayed as a Leaflet plot
}
\description{
Create Summary Map of Sampling Locations
}
