% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Generate a Water Quality Report for an organization}
\usage{
create_report(org = "TURTLEMT", extFile = "script_generateReport.qmd")
}
\arguments{
\item{org}{organization (short name in WQP) whose data should be used to build the report}

\item{extFile}{name of report-generating script, located in the inst/extdata folder of the R8WD R package}
}
\value{
Quarto markdown document
}
\description{
Generate a Water Quality Report for an organization
}
\examples{
create_report(org = 'Tribe 1')
create_report(org = 'Tribe 2')

}
