% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Calculate Coefficient of Variation}
\usage{
cv(x)
}
\arguments{
\item{x}{a vector of two or more values}
}
\value{
a single numeric value
}
\description{
Calculate Coefficient of Variation
}
\examples{
cv(c(10, 20))
cv(c(10:30))
}
