% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinWQPData.R, R/joinWQPProfiles.R
\name{joinWQPProfiles}
\alias{joinWQPProfiles}
\title{Join WQP Measurement and Site Data}
\usage{
joinWQPProfiles(fullData = "null", sites = "null")

joinWQPProfiles(fullData = "null", sites = "null")
}
\arguments{
\item{fullData}{output of query to WQP portal using readWQPdata function in 
the dataRetrieval package}

\item{sites}{output of query to WQP portal using whatWQPsites function in the
dataRetrieval package}

\item{dat.full}{output of query to WQP portal using readWQPdata function in 
the dataRetrieval package}

\item{dat.sites}{output of query to WQP portal using whatWQPdat.sites function in the
dataRetrieval package}
}
\value{
a combined data frame of WQP measurements with corresponding sampling
        site data

a combined data frame of WQP measurements with corresponding sampling
        site data
}
\description{
This function takes data from the same WQP query and combines
             site data with measurements to map observations with physical 
             sampling locations

This function takes data from the same WQP query and combines
             site data with measurements to map observations with physical 
             sampling locations
}
