% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pa.R
\name{qc_pa}
\alias{qc_pa}
\title{Quality control screening of Purple Air data}
\usage{
qc_pa(dat = dat, sensor_agreement = 0.7, column_name = "pm2.5_atm_")
}
\arguments{
\item{dat}{dataset obtained with `get_pa()`}

\item{sensor_agreement}{required level of sensor agreement, expressed as percent difference from the lower of the two sensors.}

\item{column_name}{name of the column with data of interest.}
}
\value{
dataframe
}
\description{
Quality control screening of Purple Air data
}
\examples{
qc_pa

}
