% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create Summary Map of Sampling Locations}
\usage{
create_map(.data)
}
\arguments{
\item{.data}{full WQP data frame with site location data}
}
\value{
an interactive map of sample locations displayed as a Leaflet plot
}
\description{
Create Summary Map of Sampling Locations
}
